Remove the GCC missing-braces warning.
This is GCC bug #53119:
http://gcc.gnu.org/bugzilla/show_bug.cgi?id=53119

GCC warns when the first member in a struct is an array.
I work around it with double brace pairs, the outer for the struct and the inner
for the array.

Compiling messages:
Middlewares/Third_Party/FatFs/src/ff_gen_drv.c:34:1: warning: missing braces
around initializer [-Wmissing-braces]
 Disk_drvTypeDef  disk = {0};
 ^